Initial changes and build
