+++ .env
@@ -1,0 +1,1 @@
+MONGO_URL = 'mongodb+srv://admin:admin@cluster0.nqlscmr.mongodb.net/alt_pizza?retryWrites=true&w=majority'
